server/requirements.txt
azure-functions==1.21.3
beautifulsoup4==4.12.3
lxml==5.3.0
aiohttp==3.9.5
redis==5.2.1
